Load the database file in `registerDatastore` and `define`, instead of inside each adapter method
The issue that was being solved by loading in the adapter methods was the race conditions that could arise when using NeDB's `autoload: true` feature, because loading databases in NeDB is asynchronous.  By using an async.each loop in `registerDatastore`, and loading the database file in `define` before returning, we guarantee that the files are loaded correctly before any adapter methods are called.  This simplifies the code in the adapter methods themselves, and ensures that any loading errors (like uniqueness constraint violations when auto-migrating) are caught in `registerDatastore` instead of some random adapter method.